introduce-scheduler-request-receiver-move: Move request-ingress methods to SchedulerRequestReceiver
Mechanical cut + paste for the ``introduce-scheduler-request-receiver``
mech move.

Cut ``recv_requests`` / ``recv_limit_reached`` /
``_split_work_and_control_reqs`` (@staticmethods after prep) from
Scheduler and paste them into the ``SchedulerRequestReceiver`` class body
in ``scheduler_components/request_receiver.py``.

Drop ``@staticmethod`` decorators; simplify ``self: SchedulerRequestReceiver``
type annotation to bare ``self`` (in class context the type is implicit).
Method bodies otherwise byte-identical.

All callers updated:
  ``self.recv_requests(self.request_receiver, last_forward_mode=...)`` →
  ``self.request_receiver.recv_requests(last_forward_mode=...)``
(pure prefix transformation).

Refactor chain ID: introduce-scheduler-request-receiver-move
