+++ python/sglang/srt/multiplex/multiplexing_mixin.py
@@ -110,9 +110,7 @@ class SchedulerMultiplexMixin:
         while True:
             with torch.cuda.stream(decode_stream):
                 set_pdmux_status(False)
-                recv_reqs = self.recv_requests(
-                    self.request_receiver,
-                )
+                recv_reqs = self.request_receiver.recv_requests()
                 self.process_input_requests(recv_reqs)
 
             with torch.cuda.stream(prefill_stream):
